Changed buttons to a links
- Changed buttons to a links
- Fixes

+++ Data/Entities/ProductEntity.cs
@@ -8,8 +8,8 @@ namespace Sklepix.Data.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public int Count { get; set; }
-        public int Price { get; set; }
+        public float Count { get; set; }
+        public float Price { get; set; }
         public CategoryEntity? Category { get; set; }
         public AisleEntity? Aisle { get; set; }
         public AisleRowEntity? Row { get; set; }
